services/ingestion/requirements.txt
pydantic>=2.7
psycopg2-binary>=2.9
requests>=2.31
python-dotenv>=1.0
redis>=5.0
loguru>=0.7
python-dateutil>=2.8.2
websocket-client>=1.6.0
